ResultsPanel now outputs complete command-line specification of PairedTTester.
PairedTTester now allows complete specification of ResultMatrix at the command-line and has option to turn off header output at command-line. 

ResultMatrix has small change in setOptions(), so that rows and columns are always enumerated now when printing of the corresponding names is turned off. Also, the keys for the columns are no longer shown if enumeration of columns is turned off.

git-svn-id: https://svn.cms.waikato.ac.nz/svn/weka/trunk@11542 e0a1b77d-ad91-4216-81b1-defd5f83fa92